remove unneeded imports, refactor preference api tests
